add item-by-id component, just a container for detail or edit page but manages the fetching on id change and (un)setting the current content
