+++ src/app/modules/item/item.module.ts
@@ -5,9 +5,11 @@ import { SharedComponentsModule } from '../shared-components/shared-components.m
 import { ItemDetailsComponent } from './pages/item-details/item-details.component';
 import { ItemHeaderComponent } from './components/item-header/item-header.component';
 import { ItemEditComponent } from './pages/item-edit/item-edit.component';
+import { ItemByIdComponent } from './pages/item-by-id/item-by-id.component';
 
 @NgModule({
   declarations: [
+    ItemByIdComponent,
     ItemDetailsComponent,
     ItemHeaderComponent,
     ItemEditComponent,
