+++ src/app/modules/item/pages/item-by-id/item-by-id.component.html
@@ -1,0 +1,1 @@
+<router-outlet></router-outlet>
