+++ src/app/shared/services/nav-types.ts
@@ -20,7 +20,7 @@ export function itemDetailsRoute(item: NavItem): any[] {
   if (item.attemptId) params[attemptParamName] = item.attemptId;
   else params[parentAttemptParamName] = item.parentAttemptId;
   params[pathParamName] = item.itemPath;
-  return ['items', 'details', item.itemId, params];
+  return ['items', 'by-id', item.itemId, params, 'details'];
 }
 
 /**
